Split out encoding details to separate module
This is another step to putting all the things that aren't specific to
synchronous TelescopeState outside telescope_state.py.
