+++ katsdptelstate/rdb_writer.py
@@ -18,7 +18,8 @@ import logging
 import os
 
 from .rdb_utility import encode_len
-from .telescope_state import _ensure_binary, _display_str, TelescopeState
+from .utils import ensure_binary, display_str
+from .telescope_state import TelescopeState
 
 # Basic RDB header. First 5 bytes are the standard REDIS magic
 # Next 4 bytes store the RDB format version number (6 in this case)
@@ -130,7 +131,7 @@ class RDBWriter:
             logger.info("No keys specified - dumping entire database")
             keys = client.keys(b'*')
         for key in keys:
-            key = _ensure_binary(key)
+            key = ensure_binary(key)
             dumped_value = client.dump(key)
             try:
                 if not dumped_value:
@@ -138,7 +139,7 @@ class RDBWriter:
                 encoded_str = encode_item(key, dumped_value)
             except (ValueError, KeyError) as e:
                 self.keys_failed += 1
-                logger.error("Failed to save key %s: %s", _display_str(key), e)
+                logger.error("Failed to save key %s: %s", display_str(key), e)
                 continue
             self._fileobj.write(encoded_str)
             self.keys_written += 1
