+++ katsdptelstate/utils.py
@@ -16,6 +16,22 @@
 
 import struct
 import math
+import functools
+
+import six
+
+
+# Behave gracefully in case someone uses non-UTF-8 binary in a key on PY3
+ensure_str = functools.partial(six.ensure_str, errors='surrogateescape')
+ensure_binary = functools.partial(six.ensure_binary, errors='surrogateescape')
+
+
+def display_str(s):
+    """Return most human-readable and yet accurate version of *s*."""
+    try:
+        return '{!r}'.format(six.ensure_str(s))
+    except UnicodeDecodeError:
+        return f'{s!r}'
 
 
 def pack_query_timestamp(time, is_end, include_end=False):
